Refactored Button to use Consumer
In React Context, there are multiple ways to access the data passed from
Context to Components. One way is to call set ContextType to the
imported Context, call this.context, and assign it to the value that
you want to pass into the component. Another way is to use Consumer.

- Using Consumer:
  - Rather than assigning assigning this.context to the value to be
    passed to the component, refactor the component to use the
    Consumer prop of the Context component
  - To do so, remove contextType, wrap the component's value in
    <Context.Consumer> JSX tags, and pass a child to the consumer
    as a function with the argument value to retrieve the value from
    Context. Then use the same logic as before to change the value of
    the components
  - Logic can also be removed from the child function and placed in a
    helper method which can then be called in place of the logic